+++ src/lib/services/spinner-visibility.service.ts
@@ -20,7 +20,12 @@ export class SpinnerVisibilityService {
     constructor(private pendingInterceptorService: PendingInterceptorService) {
     }
 
+    /** @deprecated Deprecated in favor of visibilityObservable$ */
     get visibilityObservable(): Observable<boolean> {
+        return this.visibilityObservable$;
+    }
+
+    get visibilityObservable$(): Observable<boolean> {
         return this._visibilitySubject.asObservable();
     }